lab_3/lab3/pyproject.toml
[tool.poetry]
name = "lab2"
version = "0.2.0"
description = ""
authors = ["James Winegar <jameswine@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
requests = "^2.30.0"
joblib = "^1.2.0"
scipy = "^1.10.1"
scikit-learn = "1.2.2"
fastapi-cache2 = {version = "0.1.9", extras = ["redis"]}
numpy = "1.25.1"
redis = "4.6.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
httpx = "^0.24.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
fastapi-cache2 = {version = "0.1.9", extras = ["redis"]}
